Spotify Project/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
spotipy>=2.23.0
flask>=2.3.0
streamlit>=1.25.0
plotly>=5.15.0
requests>=2.31.0
python-dotenv>=1.0.0
joblib>=1.3.0
